Fix light mode appearance of project info view
The site is currently always using the dark theme (as there is no code
to switch the current theme), but the github-markdown-css package's
default offering automatically switches between light and dark through
the `@media (prefers-color-scheme)` rule.

For now, switch to the always-dark version of github-markdown-css and
(since the single-theme versions do not use CSS variables) add custom
CSS to make the background transparent again.